Function mp_starknet::crypto::commitment::calculate_commitments

pub fn calculate_commitments<T: CryptoHasher>(
    transactions: &[Transaction]
) -> (H256, (H256, u128))

Calculate the transaction commitment, the event commitment and the event count.

Arguments

transactions - The transactions of the block

Returns

The transaction commitment, the event commitment and the event count.