Function mp_starknet::fees::get_transaction_resources

pub fn get_transaction_resources<S: State + StateChanges>(
    state: &mut S,
    execute_call_info: &Option<CallInfo>,
    validate_call_info: &Option<CallInfo>,
    execution_resources: &mut ExecutionResources,
    tx_type: TxType
) -> Result<BTreeMap<String, usize>, TransactionExecutionErrorWrapper>

Gets the transaction resources.

Arguments

state - State object to get the state changes.

execute_call_info - Call info of the execution of the __execute__ entrypoint.

execution_resources - Resources used by the execution.

tx_type - Type of the transaction.

Returns

BTreeMap<String, usize> - Mapping from execution resources to the number of uses.

Error

TransactionExecutionErrorWrapper if a step of the execution resources computation fails.